feat: Add Shlef component
